Layout
Fixed headers when width decreases, the h2 top banner width should
probably be it's own size as well.

Added footer, with icons.

Changed the size of the wiki "cards" and the size of the content in
them. (Still gonna need to add the hot links in it's own div so they can
be hidden/shown)  with a simple v / ^ button for the single column
layout.

Banner image still needs to be made to change width at some point